Fix length bug in 2020-01 part 1 x86
The length of the input data was divided by 2, which is a remnant from
when I generated the input data as 16-bit values. Since they're now
expected to be 32-bit values, the divisor should be 4.

An unnecessary XOR instruction is also removed, since EDI will be fully
overwritten by a subsequent MOV from ESI.